Add files via upload
This is a compressed folder containing all the source code for the system. After downloading one should create a database called ecoproc and import the sql file called databases.sql that is in the root folder. the project uses a file called database.php to connect to a mysql database inside the file is a class called Connection which has methods and attributes necessary to communicate with database. 

The code is suitable for college projects and with few touchups it can also be used for commercial use